+++ src/hpguppi_upchan_bf_thread.c
@@ -426,10 +426,9 @@ static void *run(hashpipe_thread_args_t * args)
 
       // Skip zeroth index since the number of coarse channels is even and the center frequency is between the 2 middle channels
       for(int i=0; i<n_coarse_proc; i++){
-        //coarse_chan_freq[i] = (i-((n_chan_per_node-1)/2))*coarse_chan_band + (obsfreq*1e6);
-        coarse_chan_freq[i] = ((i + subband_idx*n_coarse_proc)-((n_chan_per_node-1)/2))*(chan_bw*1e-3) + (obsfreq*1e-3);
+        coarse_chan_freq[i] = ((i + subband_idx*n_coarse_proc)-(((double)n_chan_per_node-1)/2))*(chan_bw*1e-3) + (obsfreq*1e-3);
         // Equivalent equation //
-        //coarse_chan_freq[i] = ((i + subband_idx*n_coarse_proc)-(n_chan_per_node/2))*(chan_bw*1e-3) + (obsfreq*1e-3) + ((chan_bw/2)*1e-3);
+        //coarse_chan_freq[i] = ((i + subband_idx*n_coarse_proc)-((double)n_chan_per_node/2))*(chan_bw*1e-3) + (obsfreq*1e-3) + ((chan_bw/2)*1e-3);
       }
 
       // Number of FFT points depending on mode
